fixes and tests for device/host
left off at needing to test the persistence of device/host work.

learning as I go along writing tests, perhaps these cached objects is not the
right place for calling updates to the underlying database structures. Since
they are immutable, they technically cannot update themselves, which makes
usage of their API awkward for this.

Committing what I was on so I can push to the branch and prove that it's
alive. Will wrap this up soon.